fix(ui): truncate delay minutes toward zero in info window
Match label truncation behavior so delay readouts agree across views and
align with control action thresholds (NoDelay below 1 min magnitude).
Previous "00" format rounded -0.7 to -1, showing Expedite-style negative
while underlying state was NoDelay.

+++ source/Maestro.Wpf/Converters/MinutesConverter.cs
@@ -11,7 +11,7 @@ class MinutesConverter : IValueConverter
     {
         return value switch
         {
-            TimeSpan timeSpan => timeSpan.TotalMinutes.ToString("00"),
+            TimeSpan timeSpan => ((int)timeSpan.TotalMinutes).ToString("00"),
             DateTimeOffset dateTime => dateTime.Minute.ToString("00"),
             _ => string.Empty
         };
